CommandRegistry class implementation
A registry class that handles command loading on bot startup. By default
it'll look in the dist/commands folder for command candidates, add each
one to a category map and attempt to register them. After this process
is finished, the command registry will be able to provide the bot with
each command's execute function.