EventRegistry class implementation
The EventRegistry class serves the same purpose as the CommandRegistry
class, but for events. On startup, it'll look in the dist/events folder
for events and attempt to register them. After this process is finished,
the bot's client will have listeners bound. By default, if no
interactionCreate event is found, the EventRegistry class will register
it's own base implementation.